Added KnowledgeDB to the release command

+++ src/PainKiller.PowerCommands/PainKiller.PowerCommands.MyExampleCommands/Commands/ReleaseCommand.cs
@@ -11,23 +11,26 @@ public class ReleaseCommand : CommandBase<PowerCommandsConfiguration>
         UpdateCoreDirectory(solutionPath);
         Build(solutionPath);
 
-
+        solutionPath = SolutionFileManager.GetLocalSolutionRoot().Replace("src\\PainKiller.PowerCommands\\", "Implementations\\KnowledgeDB-Implementation");
+        UpdateCoreDirectory(solutionPath);
+        Build(solutionPath);
 
         return Ok();
     }
     private void UpdateCoreDirectory(string solutionPath)
     {
         var coreImplementationDirectory = Path.Combine(solutionPath, "Core");
-        WriteLine($"Delete {coreImplementationDirectory}");
-        Directory.Delete(coreImplementationDirectory, recursive: true);
+        WriteSuccess($"Delete {coreImplementationDirectory}");
+        if(Directory.Exists(coreImplementationDirectory)) Directory.Delete(coreImplementationDirectory, recursive: true);
 
         var coreDirectory = Path.Combine(SolutionFileManager.GetLocalSolutionRoot(), "Core");
-        WriteLine($"Copy {coreDirectory} to {coreImplementationDirectory}");
+        WriteSuccess($"Copy {coreDirectory} to {coreImplementationDirectory}");
         IOService.CopyFolder(coreDirectory, coreImplementationDirectory);
         
     }
     private void Build(string solutionPath)
     {
         ShellService.Service.Execute("dotnet", "build", solutionPath, WriteLine, "", waitForExit: true);
+        WriteSuccess($"Build {solutionPath} OK");
     }
 }
